wip: move normalizer up
