+++ src/interface/http/map-result-to-http.ts
@@ -1,4 +1,4 @@
-import { normalize } from '@/framework/normalizer/normalize';
+import { normalize } from '@/framework/normalize';
 import { createResultToHttpMapper } from '@/interface/http/result-to-http';
 
 export const mapResultToHttp = createResultToHttpMapper({ normalize: normalize });
